Fix unsigned type comparison warning

+++ src/main.cpp
@@ -295,7 +295,7 @@ int main(int argc, char *argv[])
     {
         QByteArray msg;
         QByteArray message = QByteArray("HOOK");
-        for(int i=0; i<hookPayload.size(); i++)
+        for(size_t i=0; i<hookPayload.size(); i++)
         {
             message = message + QByteArray("\n")
                               + QByteArray(hookSpec[i].data(), (int)hookSpec[i].size()) + QByteArray("\n")
